Move `SharedBuilder`'s helper functions to `ckb-shared`
